+++ Logic/API/IBallLogic.cs
@@ -1,9 +1,10 @@
 ﻿using BallSimulator.Data;
+using BallSimulator.Data.API;
 using System.ComponentModel;
 
-namespace BallSimulator.Logic;
+namespace BallSimulator.Logic.API;
 
-public interface IBallLogic: IObservable<IBallLogic>, IObserver<IBall>, INotifyPropertyChanged
+public interface IBallLogic : IObservable<IBallLogic>, IObserver<IBall>, INotifyPropertyChanged
 {
     public int Diameter { get; }
     public int Radius { get; }
