+++ Model/BallsCountValidator.cs
@@ -1,5 +1,4 @@
-﻿using BallSimulator.Logic;
-using BallSimulator.Presentation.Model.API;
+﻿using BallSimulator.Presentation.Model.API;
 
 namespace BallSimulator.Presentation.Model;
