+++ Model/ModelApi.cs
@@ -1,4 +1,4 @@
-﻿using BallSimulator.Logic;
+﻿using BallSimulator.Logic.API;
 using BallSimulator.Presentation.Model.API;
 
 namespace BallSimulator.Presentation.Model;
