+++ ModelTests/ModelApiTest.cs
@@ -1,5 +1,5 @@
+using BallSimulator.Presentation.Model.API;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Model.API;
 
 namespace ModelTests
 {
