+++ Tests/BallTest.cs
@@ -1,4 +1,3 @@
-using BallSimulator.Logic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace BallSimulator.Tests
