+++ Tests/SimulationControllerTest.cs
@@ -1,5 +1,4 @@
 ﻿using BallSimulator.Data;
-using BallSimulator.Logic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
 using System.Linq;
