+++ Tests/Vector2Test.cs
@@ -1,5 +1,4 @@
-﻿using BallSimulator.Logic;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace BallSimulator.Tests
 {
